+++ gennav/envs/__init__.py
@@ -1,0 +1,1 @@
+from gennav.envs.base import Environment  # noqa: F401
