+++ gennav/planners/__init__.py
@@ -1,3 +1,4 @@
+from gennav.planners.base import Planner  # noqa: F401
 from gennav.planners.prm import PRM  # noqa: F401
 from gennav.planners.rrt import RRT  # noqa: F401
-from gennav.planners.rrtstar import RRTSTAR  # noqa: F401
+from gennav.planners.rrt import RRTStar  # noqa: F401
