+++ gennav/planners/graph_search/__init__.py
@@ -1,0 +1,1 @@
+from gennav.planners.graph_search.astar import astar  # noqa: F401
