--- gennav/planners/graph_search_algorithms/__init__.py
@@ -1,1 +1,0 @@
-from gennav.planners.graph_search_algorithms.astar import astar  # noqa: F401
