+++ gennav/planners/prm/__init__.py
@@ -1,0 +1,1 @@
+from gennav.planners.prm.prm import PRM  # noqa: F401
