+++ gennav/planners/rrt/rrt.py
@@ -1,9 +1,10 @@
 import math
 
 from descartes import PolygonPatch
-from gennav.utils.planner import check_intersection
 from matplotlib import pyplot as plt
 from shapely.geometry import Point, Polygon
+
+from gennav.utils.planner import check_intersection
 
 # from gennav.planners.samplers.samplers import uniform_adjustable_random_sampler as sampler
