+++ gennav/planners/rrt/rrt_star.py
@@ -2,6 +2,7 @@ import math
 
 import numpy as np
 from matplotlib import pyplot as plt
+
 from utils_scan import adjustable_random_sampler as sampler
 from utils_scan import (
     check_intersection_scan,
