--- gennav/planners/samplers/__init__.py
@@ -1,4 +1,0 @@
-from gennav.planners.samplers.samplers import (  # noqa: F401
-    uniform_adjustable_random_sampler,
-    uniform_random_sampler,
-)
